use context provider instead with provider and refactor app file
