+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/santiagorios/Desktop/sr-react/src/routes/routes.ts":"1","/Users/santiagorios/Desktop/sr-react/src/routes/protected/index.tsx":"2","/Users/santiagorios/Desktop/sr-react/src/routes/public/index.tsx":"3","/Users/santiagorios/Desktop/sr-react/src/components/App/index.tsx":"4","/Users/santiagorios/Desktop/sr-react/src/components/I18n/index.tsx":"5","/Users/santiagorios/Desktop/sr-react/src/screens/Home/index.tsx":"6","/Users/santiagorios/Desktop/sr-react/src/screens/Login/index.tsx":"7","/Users/santiagorios/Desktop/sr-react/src/routes/index.tsx":"8"},{"size":607,"mtime":1614883771081,"results":"9","hashOfConfig":"10"},{"size":494,"mtime":1614883871606,"results":"11","hashOfConfig":"10"},{"size":487,"mtime":1614883873776,"results":"12","hashOfConfig":"10"},{"size":1038,"mtime":1614893806099,"results":"13","hashOfConfig":"10"},{"size":1243,"mtime":1614885545810,"results":"14","hashOfConfig":"10"},{"size":967,"mtime":1614885569815,"results":"15","hashOfConfig":"10"},{"size":3270,"mtime":1614885587108,"results":"16","hashOfConfig":"10"},{"size":893,"mtime":1614883503615,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1xg9b1i",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/santiagorios/Desktop/sr-react/src/routes/routes.ts",[],"/Users/santiagorios/Desktop/sr-react/src/routes/protected/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/routes/public/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/components/App/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/components/I18n/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/screens/Home/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/screens/Login/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/routes/index.tsx",[]]
