+++ src/components/App/index.tsx
@@ -1,17 +1,15 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useReducer } from 'react';
 import i18next from 'i18next';
 
-import { Context, reducer, INITIAL_STATE, useSelector, useDispatch, actionCreators } from 'contexts';
+import { Context, reducer, INITIAL_STATE, actionCreators } from 'contexts/UserContext';
 import Routes from 'routes';
 import I18n from 'components/I18n';
-import withProvider from 'components/ProviderWrapper';
 import LocalStorageService from 'services/LocalStorageService';
 
 import styles from './styles.module.scss';
 
 function App() {
-  const language = useSelector(state => state.language);
-  const dispatch = useDispatch();
+  const [state, dispatch] = useReducer(reducer, INITIAL_STATE);
   useEffect(() => {
     const lang: string = LocalStorageService.getValue('lang');
     if (lang) {
@@ -23,14 +21,16 @@ function App() {
     } else {
       dispatch(actionCreators.logout());
     }
-  }, [language, dispatch]);
+  }, [state.language, dispatch]);
 
   return (
-    <div className={`${styles.appContainer} row center middle`}>
-      <I18n />
-      <Routes />
-    </div>
+    <Context.Provider value={{ state, dispatch }}>
+      <div className={`${styles.appContainer} row center middle`}>
+        <I18n />
+        <Routes />
+      </div>
+    </Context.Provider>
   );
 }
 
-export default withProvider({ context: Context, reducer, initialState: INITIAL_STATE })(App);
+export default App;
